# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SABABUKA Bersinar API
  version: 1.0.0-draft
  summary: Kontrak API MVP untuk tata kelola data Kabupaten Kapuas
  description: |
    API tunggal untuk ruang Superadmin, BAPPERIDA, Kominfo, OPD, Pimpinan,
    dashboard eksekutif, dan Asisten AI. Semua operasi dibatasi oleh permission,
    lingkup organisasi, status workflow, dan feature flag.
  contact:
    name: Pengelola SABABUKA Bersinar
  license:
    name: Penggunaan Internal Pemerintah Kabupaten Kapuas
    identifier: LicenseRef-SABABUKA-Internal
servers:
  - url: http://127.0.0.1:3001/api/v1
    description: Local development (aktif saat ini)
  - url: https://dev.sababuka.com/api/v1
    description: Development (belum diaktifkan)
  - url: https://sababuka.com/api/v1
    description: Production (setelah disetujui)
...
security:
  - sessionCookie: []
paths:
...
  /policy-focuses:
    get:
      tags: [Governance]
      operationId: listPolicyFocuses
      summary: Daftar fokus kebijakan
      description: Membutuhkan `policy_focus.view`.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Daftar fokus
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PolicyFocusPage"}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      tags: [Governance]
      operationId: createPolicyFocus
      summary: Membuat fokus kebijakan
      description: Membutuhkan `policy_focus.manage`.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PolicyFocusInput"}
      responses:
        "201":
          description: Fokus dibuat
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PolicyFocus"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /categories:
    get:
      tags: [Governance]
      operationId: listCategories
      summary: Daftar kategori dinamis
      description: Membutuhkan `category.view`.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - name: policy_focus_id
          in: query
          schema: {$ref: "#/components/schemas/Uuid"}
      responses:
        "200":
          description: Daftar kategori
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CategoryPage"}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      tags: [Governance]
      operationId: createCategory
      summary: Membuat kategori
      description: Membutuhkan `category.manage`.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CategoryInput"}
      responses:
        "201":
          description: Kategori dibuat
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Category"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /units:
    get:
      tags: [Governance]
      operationId: listUnits
      summary: Daftar satuan indikator aktif
      description: Membutuhkan `indicator.view`.
      responses:
        "200":
          description: Daftar satuan
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: {$ref: "#/components/schemas/Unit"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /periods:
    get:
      tags: [Governance]
      operationId: listPeriods
      summary: Daftar periode pelaporan
      description: Membutuhkan `indicator.view`.
      responses:
        "200":
          description: Daftar periode
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: {$ref: "#/components/schemas/Period"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /indicators:
    get:
      tags: [Governance]
      operationId: listIndicators
      summary: Daftar indikator
      description: Membutuhkan `indicator.view`.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - name: category_id
          in: query
          schema: {$ref: "#/components/schemas/Uuid"}
        - name: organization_id
          in: query
          schema: {$ref: "#/components/schemas/Uuid"}
        - name: status
          in: query
          schema: {$ref: "#/components/schemas/IndicatorStatus"}
      responses:
        "200":
          description: Daftar indikator
          content:
            application/json:
              schema: {$ref: "#/components/schemas/IndicatorPage"}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      tags: [Governance]
      operationId: createIndicator
      summary: Membuat indikator dan versi draft pertama
      description: Membutuhkan `indicator.manage`.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/IndicatorCreateInput"}
      responses:
        "201":
          description: Indikator dan versi draft dibuat
          content:
            application/json:
              schema: {$ref: "#/components/schemas/IndicatorDetail"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
  /indicators/{indicator_id}:
    patch:
      tags: [Governance]
      operationId: updateIndicatorDraft
      summary: Mengubah metadata versi indikator draft
      description: Membutuhkan `indicator.manage`; versi selain draft ditolak agar histori definisi tetap terjaga.
      parameters:
        - name: indicator_id
          in: path
          required: true
          schema: {$ref: "#/components/schemas/Uuid"}
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/IndicatorCreateInput"}
      responses:
        "200":
          description: Draft indikator diperbarui
          content:
            application/json:
              schema: {$ref: "#/components/schemas/IndicatorDetail"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /indicator-versions/{version_id}/actions/{action}:
    post:
      tags: [Governance]
      operationId: transitionIndicatorVersion
      summary: Menjalankan transisi versi indikator
      description: |
        Aksi `submit`, `approve`, `activate`, atau `retire`. Permission dan status
        asal diperiksa sesuai dokumen workflow MVP v1.
      parameters:
        - $ref: "#/components/parameters/VersionId"
        - name: action
          in: path
          required: true
          schema:
            type: string
            enum: [submit, approve, activate, retire]
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/TransitionInput"}
      responses:
        "200":
          description: Transisi berhasil
          content:
            application/json:
              schema: {$ref: "#/components/schemas/IndicatorVersion"}
        "400": {$ref: "#/components/responses/InvalidTransition"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
...
  /audit/events:
    get:
      tags: [Audit]
      operationId: listAuditEvents
      summary: Membaca audit sesuai scope
      description: Membutuhkan `audit.view`.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: event_type
          in: query
          schema: {type: string}
        - name: entity_type
          in: query
          schema: {type: string}
        - name: actor_id
          in: query
          schema: {$ref: "#/components/schemas/Uuid"}
        - name: occurred_from
          in: query
          schema: {type: string, format: date-time}
        - name: occurred_until
          in: query
          schema: {type: string, format: date-time}
      responses:
        "200":
          description: Daftar audit append-only
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AuditPage"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /notifications:
    get:
      tags: [System]
      operationId: listNotifications
      summary: Daftar notifikasi milik pengguna aktif
      parameters:
        - name: unread_only
          in: query
          schema: {type: boolean, default: false}
      responses:
        "200": {description: Daftar notifikasi dan jumlah belum dibaca}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /notifications/{notification_id}/read:
    post:
      tags: [System]
      operationId: markNotificationRead
      summary: Menandai notifikasi milik pengguna sebagai dibaca
      parameters:
        - name: notification_id
          in: path
          required: true
          schema: {$ref: "#/components/schemas/Uuid"}
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204": {description: Notifikasi ditandai dibaca}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
  /notifications/read-all:
    post:
      tags: [System]
      operationId: markAllNotificationsRead
      summary: Menandai seluruh notifikasi pengguna sebagai dibaca
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204": {description: Seluruh notifikasi ditandai dibaca}
        "401": {$ref: "#/components/responses/Unauthorized"}
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sababuka_session
      description: Sesi server-side; cookie wajib HttpOnly, Secure, dan SameSite=Lax.
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string, minLength: 20}
      description: Token CSRF terikat ke sesi untuk seluruh operasi mutasi.
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    Search:
      name: q
      in: query
      schema: {type: string, maxLength: 200}
...
    VersionId:
      name: version_id
      in: path
      required: true
      schema: {$ref: "#/components/schemas/Uuid"}
...
  responses:
    Unauthorized:
      description: Sesi tidak tersedia atau kedaluwarsa
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
          example:
            error: {code: AUTH_REQUIRED, message: Sesi diperlukan., request_id: 8be45d40-e437-4f44-86da-91d0cae75488}
    Forbidden:
      description: Permission atau scope tidak mencukupi
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    NotFound:
      description: Objek tidak ditemukan atau tidak boleh diungkapkan
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    ValidationError:
      description: Permintaan tidak valid
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    InvalidTransition:
      description: Transisi workflow tidak valid
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
...
    Conflict:
      description: Konflik kode, versi, atau data aktif
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    PageMeta:
      type: object
      required: [page, page_size, total_items, total_pages]
      properties:
        page: {type: integer, minimum: 1}
        page_size: {type: integer, minimum: 1}
        total_items: {type: integer, minimum: 0}
        total_pages: {type: integer, minimum: 0}
    PageEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {type: array, items: {}}
        meta: {$ref: "#/components/schemas/PageMeta"}
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {$ref: "#/components/schemas/Uuid"}
            details:
              type: array
              items:
                type: object
                additionalProperties: true
...
    PolicyFocusInput:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code: {type: string, maxLength: 64}
        name: {type: string, maxLength: 255}
        description: {type: [string, "null"]}
        display_order: {type: integer, default: 0}
    PolicyFocus:
      allOf:
        - $ref: "#/components/schemas/PolicyFocusInput"
        - type: object
          required: [id, is_active]
          properties:
            id: {$ref: "#/components/schemas/Uuid"}
            is_active: {type: boolean}
    PolicyFocusPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            data:
              type: array
              items: {$ref: "#/components/schemas/PolicyFocus"}
    CategoryInput:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code: {type: string, maxLength: 64}
        name: {type: string, maxLength: 255}
        description: {type: [string, "null"]}
        policy_focus_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        parent_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        display_order: {type: integer, default: 0}
    Category:
      allOf:
        - $ref: "#/components/schemas/CategoryInput"
        - type: object
          required: [id, is_active]
          properties:
            id: {$ref: "#/components/schemas/Uuid"}
            is_active: {type: boolean}
            policy_focus_name: {type: [string, "null"]}
            indicator_count: {type: integer, minimum: 0}
    CategoryPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            data:
              type: array
              items: {$ref: "#/components/schemas/Category"}
    Unit:
      type: object
      required: [id, code, name, decimal_places]
      properties:
        id: {$ref: "#/components/schemas/Uuid"}
        code: {type: string}
        name: {type: string}
        symbol: {type: [string, "null"]}
        decimal_places: {type: integer, minimum: 0, maximum: 10}
    Period:
      type: object
      required: [id, code, label, period_type, starts_on, ends_on]
      properties:
        id: {$ref: "#/components/schemas/Uuid"}
        code: {type: string}
        label: {type: string}
        period_type: {type: string, enum: [annual, semester, quarter, monthly, custom]}
        starts_on: {type: string, format: date}
        ends_on: {type: string, format: date}
    IndicatorStatus:
      type: string
      enum: [draft, in_review, approved, active, retired]
    IndicatorCreateInput:
      type: object
      additionalProperties: false
      required: [code, name, category_id, definition, unit_id, frequency, data_type, effective_from]
      properties:
        code: {type: string, maxLength: 80}
        name: {type: string, maxLength: 255}
        category_id: {$ref: "#/components/schemas/Uuid"}
        owner_organization_id: {$ref: "#/components/schemas/Uuid"}
        definition: {type: string, minLength: 1}
        formula: {type: [string, "null"]}
        unit_id: {$ref: "#/components/schemas/Uuid"}
        frequency: {type: string, enum: [annual, semester, quarter, monthly, event, custom]}
        data_type: {type: string, enum: [number, integer, percentage, currency, text, boolean]}
        direction:
          type: [string, "null"]
          enum: [increase, decrease, maintain, null]
        source_reference: {type: [string, "null"], maxLength: 4000}
        dimension_schema: {type: object, additionalProperties: true}
        access_level: {type: string, enum: [public, internal, restricted], default: internal}
        effective_from: {type: string, format: date}
        organizations:
          type: array
          items:
            type: object
            required: [organization_id, responsibility]
            properties:
              organization_id: {$ref: "#/components/schemas/Uuid"}
              responsibility: {type: string, enum: [primary_producer, supporter, validator, curator]}
              is_primary: {type: boolean, default: false}
        targets:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [period_id]
            properties:
              period_id: {$ref: "#/components/schemas/Uuid"}
              numeric_value: {type: [number, "null"]}
              text_value: {type: [string, "null"]}
              notes: {type: [string, "null"]}
    IndicatorVersion:
      type: object
      required: [id, version_number, definition, unit_id, frequency, data_type, status, effective_from]
      properties:
        id: {$ref: "#/components/schemas/Uuid"}
        version_number: {type: integer, minimum: 1}
        definition: {type: string}
        formula: {type: [string, "null"]}
        unit_id: {$ref: "#/components/schemas/Uuid"}
        frequency: {type: string}
        data_type: {type: string}
        direction:
          type: [string, "null"]
          enum: [increase, decrease, maintain, null]
        source_reference: {type: [string, "null"]}
        dimension_schema: {type: object, additionalProperties: true}
        access_level: {type: string}
        status: {$ref: "#/components/schemas/IndicatorStatus"}
        effective_from: {type: string, format: date}
    IndicatorDetail:
      type: object
      required: [id, code, name, category_id, category_name, version_id, version_number, definition, unit_id, unit_name, status, organizations, targets]
      properties:
        id: {$ref: "#/components/schemas/Uuid"}
        code: {type: string}
        name: {type: string}
        category_id: {$ref: "#/components/schemas/Uuid"}
        owner_organization_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        category_name: {type: string}
        owner_organization_name: {type: [string, "null"]}
        version_id: {$ref: "#/components/schemas/Uuid"}
        version_number: {type: integer, minimum: 1}
        definition: {type: string}
        formula: {type: [string, "null"]}
        unit_id: {$ref: "#/components/schemas/Uuid"}
        unit_name: {type: string}
        unit_symbol: {type: [string, "null"]}
        frequency: {type: string}
        data_type: {type: string}
        direction:
          type: [string, "null"]
          enum: [increase, decrease, maintain, null]
        source_reference: {type: [string, "null"]}
        access_level: {type: string}
        effective_from: {type: string, format: date}
        status: {$ref: "#/components/schemas/IndicatorStatus"}
        organizations:
          type: array
          items: {type: object, additionalProperties: true}
        targets:
          type: array
          items: {type: object, additionalProperties: true}
    IndicatorPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            data:
              type: array
              items: {$ref: "#/components/schemas/IndicatorDetail"}
...
    TransitionInput:
      type: object
      additionalProperties: false
      properties:
        notes: {type: [string, "null"], maxLength: 4000}
...
    AuditEvent:
      type: object
      required: [id, event_type, entity_type, occurred_at]
      properties:
        id: {$ref: "#/components/schemas/Uuid"}
        actor_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        event_type: {type: string}
        entity_type: {type: string}
        entity_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        metadata: {type: object, additionalProperties: true}
        request_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        occurred_at: {$ref: "#/components/schemas/Timestamp"}
    AuditPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            data:
              type: array
              items: {$ref: "#/components/schemas/AuditEvent"}